Problem : App architecture introduces inefficiencies and overhead
Solution: removed test and stylesheet folders. removed sendMessage from methods.  #69
