+++ imports/ui/pages/signin/signin.scss
@@ -1,4 +1,4 @@
-@import '../../stylesheets/variables';
+@import '../../layouts/variables';
 
 .sign-in{
   position: relative;
